Add files via upload
This repository contains the solutions and notes from my 3-month journey mastering Data Structures and Algorithms (DSA) with Java. I completed this journey on 31st September 2024, following along with tutorials from FreeCodeCamp via YouTube.

In this repository, you'll find:

    Solutions to various DSA problems.
    Code implementations for key algorithms.
    Data structures like arrays, linked lists, stacks, queues, trees, and more.
    Problem-solving techniques and optimizations in Java.

Feel free to explore, learn, and contribute! This repository is a reflection of my journey and a resource for anyone looking to strengthen their DSA skills in Java.

#DSA #Java #Algorithms #DataStructures #FreeCodeCamp #CodingJourney #ProblemSolving